feat: extract and clean up of view card
