+++ src/ui/features/viewer/components/scene-manager/view-card/CameraStatus.tsx
@@ -19,6 +19,7 @@ export function CameraStatus(props: CameraStatusProps) {
         }
     }
 
+    // Render the component.
     return (
         <div
             title={statusText}
